+++ application/views/fleet/plane_item.php
@@ -7,7 +7,7 @@
                     <td colspan="2"><h1>Duck Duck Goose</h1></td>
                 </tr>
                 <tr>
-                    <td colspan="2"><h2>Planes info</h2></td>
+                    <td colspan="2"><h2>Plane Info</h2></td>
                 </tr>
                 <tr>
                     <td>Name:</td>